requirements.txt
bit>=0.7.0
python-telegram-bot